put logger in util package, move stuff out of analysis.util to avoid name conflicts, put bitvector eval in its own object
